refactor files into proper folders
