+++ client/src/components/loginForm/LoginForm.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
-import "../css/LoginForm.css";
+import "../../css/LoginForm.css";
 import { Link } from "react-router-dom";
-import { GoogleCredentialResponse } from "../types";
+import { GoogleCredentialResponse } from "../../types/types";
 
 const LoginForm = () => {
   const clientId = process.env.REACT_APP_GOOGLE_CLIENT_ID;
