--- client/src/types.ts
@@ -1,3 +1,0 @@
-export type GoogleCredentialResponse = {
-  credential: string;
-};
