+++ client/src/types/types.ts
@@ -1,4 +1,8 @@
 export type Account = {
-    username: string;
-    password: string;
- }
+  username: string;
+  password: string;
+}
+
+export type GoogleCredentialResponse = {
+  credential: string;
+};
