+++ client/src/views/LoginHomepage.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "../components/LoginForm";
+import LoginForm from "../components/loginForm/LoginForm";
 import "../css/LoginHomepage.css";
 
 export const LoginHomepage = () => {
